Revert "Amend review"
