+++ backend/review/permissions.py
@@ -6,4 +6,5 @@ class IsUserReview(BasePermission):
     def has_object_permission(self, request, view, obj):
         user = request.user
         profile = Profile.objects.get(user=user)
+
         return obj.student == StudentProfile.objects.get(profile=profile)
