+++ backend/review/utils.py
@@ -1,8 +1,7 @@
 from reservation.models import Reservation
 from schedule.models import Schedule
 from review.models import Review
-from profile.models import LecturerProfile
-from django.db.models import F, Prefetch
+from django.db.models import F
 from datetime import datetime, timedelta
 from django.utils.timezone import make_aware
 from pytz import timezone, utc
@@ -23,59 +22,45 @@ def remind_review():
             diff__lte=timedelta(hours=REMINDER_TIME * 2),
             diff__gt=timedelta(hours=REMINDER_TIME),
         )
-        .prefetch_related(Prefetch("lecturer", queryset=LecturerProfile.objects.all()))
     )
-
-    reservations = Reservation.objects.filter(schedule__in=schedules).select_related(
-        "student", "lesson"
-    )
-
-    reservation_dict = {}
-    for reservation in reservations:
-        reservation_dict.setdefault(reservation.schedule.id, []).append(reservation)
 
     for schedule in schedules:
-        for reservation in reservation_dict.get(schedule.id, []):
+        reservations = Reservation.objects.filter(schedule=schedule)
+
+        for reservation in reservations:
             student = reservation.student
             lesson = reservation.lesson
 
-            review_exists = Review.objects.filter(
+            review = Review.objects.filter(
                 student=student, lesson=lesson, lecturer=schedule.lecturer
-            ).exists()
+            )
 
-            if not review_exists:
+            if not review.exists():
                 start_time = (
                     schedule.start_time.replace(tzinfo=utc)
                     .astimezone(timezone("Europe/Warsaw"))
                     .strftime("%d-%m-%Y %H:%M")
                 )
-                lecturer_full_name = schedule.lecturer.full_name
-
+                lecturer_full_name = f"{schedule.lecturer.profile.user.first_name} {schedule.lecturer.profile.user.last_name}"
+                # send reminder
                 data = {
-                    "lesson_title": schedule.lesson.title,
-                    "lecturer_full_name": lecturer_full_name,
-                    "lesson_start_time": start_time,
+                    **{
+                        "lesson_title": schedule.lesson.title,
+                        "lecturer_full_name": lecturer_full_name,
+                        "lesson_start_time": start_time,
+                    }
                 }
-
                 mailer.send(
                     email_template="review_reminder.html",
-                    to=[student.profile.user.email],
+                    to=[reservation.student.profile.user.email],
                     subject="Prośba o ocenę szkolenia",
                     data=data,
                 )
-
                 notify(
-                    profile=student.profile,
+                    profile=reservation.student.profile,
                     title="Prośba o ocenę szkolenia",
                     subtitle=schedule.lesson.title,
-                    description=(
-                        f"Proszę daj nam znać jak nam poszło. "
-                        f"Dodaj recenzję lekcji, która odbyła się {start_time} (PL) "
-                        f"i była prowadzona przez {lecturer_full_name}."
-                    ),
-                    path=(
-                        f"/account/reviews?review_status_exclude=brak&page_size=10&"
-                        f"lesson_title={quote_plus(schedule.lesson.title)}"
-                    ),
+                    description=f"Proszę daj nam znać jak nam poszło. Dodaj recenzję lekcji, która odbyła się {start_time} (PL) i była prowadzona przez {lecturer_full_name}.",
+                    path=f"/account/reviews?review_status_exclude=brak&page_size=10&lesson_title={quote_plus(schedule.lesson.title)}",
                     icon="mdi:rate-review",
                 )
