+++ backend/tests/test_ordering.py
@@ -403,6 +403,278 @@ class CourseOrderTest(APITestCase):
                 self.assertEqual(field_values, sorted(field_values, reverse=True))
 
 
+class ReviewOrderTest(APITestCase):
+    def setUp(self):
+        self.endpoint = "/api/reviews"
+        self.data = {
+            "email": "test_email@example.com",
+            "password": "TestPassword123",
+        }
+        self.user = create_user(
+            first_name="first_name",
+            last_name="last_name",
+            email=self.data["email"],
+            password=self.data["password"],
+            is_active=True,
+        )
+        self.profile = create_student_profile(profile=create_profile(user=self.user))
+        self.user_2 = create_user(
+            first_name="first_name",
+            last_name="last_name",
+            email="test2@example.com",
+            password="Test12345",
+            is_active=True,
+        )
+        self.profile_2 = create_student_profile(
+            profile=create_profile(user=self.user_2)
+        )
+        self.lecturer_user_1 = create_user(
+            first_name="first_name",
+            last_name="last_name",
+            email="lecturer_1@example.com",
+            password=self.data["password"],
+            is_active=True,
+        )
+        self.lecturer_user_2 = create_user(
+            first_name="first_name",
+            last_name="last_name",
+            email="lecturer_2@example.com",
+            password=self.data["password"],
+            is_active=True,
+        )
+        self.lecturer_profile_1 = create_lecturer_profile(
+            profile=create_profile(user=self.lecturer_user_1, user_type="W")
+        )
+        self.lecturer_profile_2 = create_lecturer_profile(
+            profile=create_profile(user=self.lecturer_user_2, user_type="W")
+        )
+
+        self.technology_1 = create_technology(name="Python")
+        self.technology_2 = create_technology(name="JS")
+        self.technology_3 = create_technology(name="VBA")
+
+        # course 1
+        self.lesson_1 = create_lesson(
+            title="Python lesson 1",
+            description="bbbb",
+            duration="90",
+            github_url="https://github.com/loopedupl/lesson",
+            price="9.99",
+            technologies=[self.technology_1],
+        )
+        self.lesson_2 = create_lesson(
+            title="Python lesson 2",
+            description="bbbb",
+            duration="30",
+            github_url="https://github.com/loopedupl/lesson",
+            price="2.99",
+            technologies=[self.technology_1],
+        )
+
+        self.topic_1 = create_topic(name="You will learn how to code")
+        self.topic_2 = create_topic(name="You will learn a new IDE")
+
+        self.skill_1 = create_skill(name="coding")
+        self.skill_2 = create_skill(name="IDE")
+
+        self.module_1 = create_module(
+            title="Module 1", lessons=[self.lesson_1, self.lesson_2]
+        )
+
+        self.course_1 = create_course(
+            title="Python Beginner",
+            description="Learn Python today",
+            level="Podstawowy",
+            skills=[self.skill_1, self.skill_2],
+            topics=[
+                self.topic_1,
+                self.topic_2,
+            ],
+            modules=[self.module_1],
+        )
+
+        self.review_course_1_1 = create_review(
+            lesson=self.lesson_1,
+            student=self.profile,
+            lecturer=self.lecturer_profile_1,
+            rating=5,
+            review="Great lesson.",
+        )
+        self.review_course_1_2 = create_review(
+            lesson=self.lesson_1,
+            student=self.profile_2,
+            lecturer=self.lecturer_profile_1,
+            rating=4,
+            review="Good lesson.",
+        )
+        self.review_course_1_3 = create_review(
+            lesson=self.lesson_2,
+            student=self.profile,
+            lecturer=self.lecturer_profile_1,
+            rating=3,
+            review="So so lesson.",
+        )
+
+        # course 2
+        self.lesson_3 = create_lesson(
+            title="JS lesson 1",
+            description="bbbb",
+            duration="90",
+            github_url="https://github.com/loopedupl/lesson",
+            price="9.99",
+            technologies=[self.technology_2],
+        )
+        self.lesson_4 = create_lesson(
+            title="JS lesson 2",
+            description="bbbb",
+            duration="30",
+            github_url="https://github.com/loopedupl/lesson",
+            price="2.99",
+            technologies=[self.technology_2],
+        )
+        self.lesson_5 = create_lesson(
+            title="JS lesson 3",
+            description="bbbb",
+            duration="120",
+            github_url="https://github.com/loopedupl/lesson",
+            price="2.99",
+            technologies=[self.technology_2],
+        )
+        self.module_2 = create_module(
+            title="Module 2", lessons=[self.lesson_3, self.lesson_4, self.lesson_5]
+        )
+
+        self.course_2 = create_course(
+            title="Javascript course for Advanced",
+            description="Course for programmers",
+            level="Zaawansowany",
+            skills=[self.skill_1, self.skill_2],
+            topics=[
+                self.topic_1,
+                self.topic_2,
+            ],
+            modules=[self.module_2],
+        )
+
+        self.review_course_2_1 = create_review(
+            lesson=self.lesson_3,
+            student=self.profile,
+            lecturer=self.lecturer_profile_1,
+            rating=5,
+            review="Great lesson.",
+        )
+        self.review_course_2_2 = create_review(
+            lesson=self.lesson_4,
+            student=self.profile,
+            lecturer=self.lecturer_profile_1,
+            rating=4,
+            review="Good lesson.",
+        )
+        self.review_course_2_3 = create_review(
+            lesson=self.lesson_3,
+            student=self.profile_2,
+            lecturer=self.lecturer_profile_1,
+            rating=2,
+            review="So so lesson.",
+        )
+        self.review_course_2_4 = create_review(
+            lesson=self.lesson_4,
+            student=self.profile_2,
+            lecturer=self.lecturer_profile_1,
+            rating=1,
+            review="So so lesson.",
+        )
+        self.review_course_2_5 = create_review(
+            lesson=self.lesson_5,
+            student=self.profile_2,
+            lecturer=self.lecturer_profile_1,
+            rating=4,
+            review="So so lesson.",
+        )
+
+        # course 3
+        self.lesson_6 = create_lesson(
+            title="VBA lesson 1",
+            description="bbbb",
+            duration="90",
+            github_url="https://github.com/loopedupl/lesson",
+            price="9.99",
+            technologies=[self.technology_3],
+        )
+        self.module_3 = create_module(title="Module 3", lessons=[self.lesson_6])
+
+        self.course_3 = create_course(
+            title="VBA course for Expert",
+            description="Course for programmers",
+            level="Ekspert",
+            skills=[self.skill_1, self.skill_2],
+            topics=[
+                self.topic_1,
+                self.topic_2,
+            ],
+            modules=[self.module_3],
+        )
+
+        self.review_course_3_1 = create_review(
+            lesson=self.lesson_6,
+            student=self.profile,
+            lecturer=self.lecturer_profile_1,
+            rating=5,
+            review="Great lesson.",
+        )
+        self.review_course_3_2 = create_review(
+            lesson=self.lesson_6,
+            student=self.profile_2,
+            lecturer=self.lecturer_profile_1,
+            rating=2,
+            review="So so lesson.",
+        )
+
+        self.fields = ["created_at"]
+
+    def test_ordering(self):
+        for field in self.fields:
+            # no login
+            self.assertFalse(auth.get_user(self.client).is_authenticated)
+            # get data
+            response = self.client.get(f"{self.endpoint}?sort_by={field}")
+            self.assertEqual(response.status_code, status.HTTP_200_OK)
+            data = json.loads(response.content)
+            count = data["records_count"]
+            results = data["results"]
+            self.assertEqual(count, 10)
+            field_values = [course[field] for course in results]
+            if isinstance(field_values[0], dict):
+                self.assertEqual(
+                    field_values, sorted(field_values, key=lambda d: d["name"])
+                )
+            else:
+                field_values = [
+                    field_value if not is_float(field_value) else float(field_value)
+                    for field_value in field_values
+                ]
+                self.assertEqual(field_values, sorted(field_values))
+            # get data
+            response = self.client.get(f"{self.endpoint}?sort_by=-{field}")
+            self.assertEqual(response.status_code, status.HTTP_200_OK)
+            data = json.loads(response.content)
+            count = data["records_count"]
+            results = data["results"]
+            self.assertEqual(count, 10)
+            field_values = [course[field] for course in results]
+            if isinstance(field_values[0], dict):
+                self.assertEqual(
+                    field_values,
+                    sorted(field_values, key=lambda d: d["name"], reverse=True),
+                )
+            else:
+                field_values = [
+                    field_value if not is_float(field_value) else float(field_value)
+                    for field_value in field_values
+                ]
+                self.assertEqual(field_values, sorted(field_values, reverse=True))
+
+
 class ScheduleOrderTest(APITestCase):
     def setUp(self):
         self.endpoint = "/api/schedules"
